refactor: изменит имя типа store -> state
